runtime: adding a hook to convert certain blocks to use in-place buffers.
If a block has history=1, relative_rate=1, 1 input, and 1 output, it is converted to an in-place buffer. The buffer object is created as usual but instead of calling the circbuf factory, we pass it the base address to point to from the upstream buffer.

When connecting blocks, each buffer in a chain of in-place blocks must have a reader attached to all upstream buffers. This includes the non in-place block that comes at the end of the chain.

This currently mostly works with; I think I'm missing a reader somewhere as my output can be off by 1 on occarion.

NOTE: this cannot be used with GRC right now. It requires that blocks are constructed in order of their connections, which is currently not guaranteed.